% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_superobj}
\alias{construct_superobj}
\title{Construct a superlda data object}
\usage{
construct_superobj(corpus_df, text = "text", labels = "labels",
  verbose = FALSE)
}
\arguments{
\item{corpus_df}{The data.table (or data.frame) used to construct the superlda data object}

\item{text}{Name of the column holding the text data.}

\item{labels}{Name of the column holding the label data.}

\item{verbose}{Print progress (useful for large datasets).}
}
\value{
superlda data object
}
\description{
Creates a superlda data object from a data.table (or data.frame) using the text2vec package. This returns a list with the following elements:
\itemize{
\item \emph{vocab} -- A character vector of the unique \emph{tokens}
\item \emph{token_vectorizer} -- The text2vec vectorizer used to generate the document-token matrix
\item \emph{dtm} -- A sparse document-token matrix (dgTMatrix)
\item \emph{labels} -- A character vector of the unique \emph{labels}
\item \emph{label_vectorizer} -- The text2vec vectorizer used to generate the document-label matrix
\item \emph{dlm} -- A sparse document-term matrix (dgTMatrix)
}
}
\examples{
# Load data table
dat = load("election_media.Rdata")

# Construct superlda data object using defaults:
slda_obj = construct_superlda(dat)
}
\author{
Travis G. Coan \href{mailto:t.coan@exeter.ac.uk}{t.coan@exeter.ac.uk}
}
