% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_label_name}
\alias{get_label_name}
\title{Return label names for a vector label indexes}
\usage{
get_label_name(label_ids, slda_obj)
}
\arguments{
\item{slda_obj}{A superlda data object (see \link{convert_superlda} for more information on superlda data objects).}

\item{token_ids}{Vector of label indices}
}
\value{
Returns a character vector of label names.
}
\description{
Returns the label names associated with a vector of label indexes.
}
\author{
Travis G. Coan \href{mailto:t.coan@exeter.ac.uk}{t.coan@exeter.ac.uk}
}
