% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_token_name}
\alias{get_token_name}
\title{Return token names for a vector token indexes}
\usage{
get_token_name(token_ids, slda_obj)
}
\arguments{
\item{token_ids}{Vector of token indices}

\item{slda_obj}{A superlda data object (see \link{convert_superlda} for more information on superlda data objects).}
}
\value{
Returns a character vector of token names.
}
\description{
Returns the token names associated with a vector of token indexes.
}
\author{
Travis G. Coan \href{mailto:t.coan@exeter.ac.uk}{t.coan@exeter.ac.uk}
}
