% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_top_keywords}
\alias{get_top_keywords}
\title{Return the most probable tokens for each label}
\usage{
get_top_keywords(word_probs, slda_obj, n = 10)
}
\arguments{
\item{word_probs}{Word (or token) probability matrix from an estimated model}

\item{slda_obj}{A superlda data object (see \link{convert_superlda} for more information on superlda data objects).}

\item{n}{Number of keywords to return (default: n = 10)}
}
\description{
Returns the "n" most probable tokens (or "keywords) for each label
}
\author{
Travis G. Coan \href{mailto:t.coan@exeter.ac.uk}{t.coan@exeter.ac.uk}
}
