% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_top_labels}
\alias{get_top_labels}
\title{Return the most probable labels for each document}
\usage{
get_top_labels(doc_probs, corpus, n = 3, rtype = "label")
}
\arguments{
\item{doc_probs}{Document probability matrix from an estimated model}

\item{n}{Number of labels to return (default: n = 3)}

\item{slda_obj}{A superlda data object (see \link{convert_superlda} for more information on superlda data objects).}
}
\description{
Returns the "n" most probable labels for each document
}
\author{
Travis G. Coan \href{mailto:t.coan@exeter.ac.uk}{t.coan@exeter.ac.uk}
}
