% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlda.R
\docType{package}
\name{superlda}
\alias{superlda}
\alias{superlda-package}
\title{superlda}
\description{
Fits supervised, generative models.
}
\author{
Travis G. Coan \href{mailto:t.coan@exeter.ac.uk}{t.coan@exeter.ac.uk}
}
