% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{transform_superobj}
\alias{transform_superobj}
\title{Make data consistent with an existing superlda data object}
\usage{
transform_superobj(corpus_df, slda_obj, text = "text")
}
\arguments{
\item{corpus_df}{The data.table (or data.frame) used to construct the superlda data object}

\item{text}{Name of the column holding the text data.}

\item{labels}{Name of the column holding the label data.}

\item{verbose}{Print progress (useful for large datasets).}
}
\value{
superlda data object
}
\description{
Transforms a data.table (or data.frame) such at the returned features are consistent with an exisitng superlda data object. This returns a list with the following elements:
\itemize{
\item \emph{vocab} -- A character vector of the unique \emph{tokens}
\item \emph{token_vectorizer} -- The text2vec vectorizer used to generate the document-token matrix
\item \emph{dtm} -- A sparse document-token matrix (dgTMatrix)
\item \emph{labels} -- A character vector of the unique \emph{labels}
\item \emph{label_vectorizer} -- The text2vec vectorizer used to generate the document-label matrix
\item \emph{dlm} -- A sparse document-term matrix (dgTMatrix)
}
}
\examples{
# Load data table
dat = load("election_media.Rdata")

# Construct superlda data object using defaults:
slda_obj = construct_superlda(dat)
}
\author{
Travis G. Coan \href{mailto:t.coan@exeter.ac.uk}{t.coan@exeter.ac.uk}
}
